Update code to be compatible with scala 2.10
This is mainly a change to the source code of scalagen itself, the largest change being the removal of the - and -- operators from List.

There is also a small change in the generated source, which requires knowledge of whether we are targetting scala 2.9 or 2.10. StaticAnnotation in scala has moved to a package that is not imported by default, so if we're targetting 2.10 we must add an explicit import.

The change to Annotations doesn't need a new test because it is already covered by ScalaCompilationTest (the annotations Config.java and QueryDelegate.java would fail to compile under 2.10 without this change).